Add files via upload
1. PHARMACY
John is the administrator of the “Smiles” Pharmacy. He needs a software application to help him manage his pharmacy's medication stocks. Each Medication is stored in the following way: name, concentration, quantity and price. John wants his new application to help him in the following ways:
a. He wants to be able to add, delete or update a medication. A
medicine is uniquely identified by its name and concentration. If a product
that already exists is added, its quantity will be updated (the new quantity is
added to the existing one).
b. He wants to be able to see all the available medications containing a given string (if the string is empty, all the available medications will be considered), sorted ascending by medication name.
c. He wants to be able to see only those medications that are in short supply (quantity less than X items, where the value of X is provided by John).
d. The application must provide multiple undo and redo functionality. Each step will undo